% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{compoundCluster}
\alias{compoundCluster}
\title{Cluster compounds in reduced space}
\usage{
compoundCluster(dimR, method = "dbscan", eps = 0.5, minPts = 5)
}
\arguments{
\item{dimR}{A \code{cellpainting_dimred} object.}

\item{method}{Clustering method. Currently only \code{"dbscan"} is supported.}

\item{eps}{DBSCAN epsilon.}

\item{minPts}{DBSCAN \code{minPts}.}
}
\value{
A \code{cellpainting_clusters} object.
}
\description{
Cluster compounds in reduced space
}
\details{
\code{Compound} is required. Rows where \code{Compound} is \code{NA} are clustered
together as unknown wells; non-missing values are clustered per compound.
}
