% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{curveProject}
\alias{curveProject}
\title{Project unknown compound clusters onto known concentration trajectory curves}
\usage{
curveProject(predCurve, newData = NULL, eps = 0.5, minPts = 5)
}
\arguments{
\item{predCurve}{A \code{cellpainting_curves} object.}

\item{newData}{Optional \code{cellpainting_dimred} object.}

\item{eps}{DBSCAN epsilon for unknown clustering.}

\item{minPts}{DBSCAN \code{minPts} for unknown clustering.}
}
\value{
A list containing all projections and per-cluster best matches.
}
\description{
Project unknown compound clusters onto known concentration trajectory curves
}
