% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{normalize}
\alias{normalize}
\title{Normalize Cell Painting features}
\usage{
normalize(data, method = "z.score", by = "Plate")
}
\arguments{
\item{data}{A \code{cellpainting_data} object.}

\item{method}{Normalization method. Currently only \code{"z.score"} is supported.}

\item{by}{Optional grouping columns used to compute normalization per group.}
}
\value{
A \code{cellpainting_normalized} object.
}
\description{
Normalize Cell Painting features
}
