% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDimred}
\alias{plotDimred}
\title{Plot reduced Cell Painting coordinates}
\usage{
plotDimred(dimR, compound_name_map = NULL)
}
\arguments{
\item{dimR}{A \code{cellpainting_dimred} object.}

\item{compound_name_map}{Optional named character vector used to recode compound labels.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot reduced Cell Painting coordinates
}
