% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotKnownClusters}
\alias{plotKnownClusters}
\title{Plot DBSCAN clusters for one known compound}
\usage{
plotKnownClusters(predClust, compound)
}
\arguments{
\item{predClust}{A \code{cellpainting_clusters} object.}

\item{compound}{Compound name to plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot DBSCAN clusters for one known compound
}
