% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotKnownCurves}
\alias{plotKnownCurves}
\title{Plot a concentration trajectory curve for one known compound}
\usage{
plotKnownCurves(predCurve, compound)
}
\arguments{
\item{predCurve}{A \code{cellpainting_curves} object.}

\item{compound}{Compound name to plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a concentration trajectory curve for one known compound
}
