% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotUnknownClusters}
\alias{plotUnknownClusters}
\title{Plot DBSCAN clusters for unknown compounds}
\usage{
plotUnknownClusters(predClust)
}
\arguments{
\item{predClust}{A \code{cellpainting_clusters} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot DBSCAN clusters for unknown compounds
}
