% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotUnknownProjections}
\alias{plotUnknownProjections}
\title{Plot unknown-cluster projections onto one known compound curve}
\usage{
plotUnknownProjections(predCurve, projections, compound)
}
\arguments{
\item{predCurve}{A \code{cellpainting_curves} object.}

\item{projections}{Projection results returned by \code{curveProject()}.}

\item{compound}{Compound name to plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot unknown-cluster projections onto one known compound curve
}
