% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{selectFeatureCluster}
\alias{selectFeatureCluster}
\title{Select cluster-discriminating features among unknown clusters}
\usage{
selectFeatureCluster(
  predClust,
  data.norm,
  method = "wilcox",
  p.adjust = "bonferroni",
  cutoff = 0.05,
  medianDiffCutoff = 0.5
)
}
\arguments{
\item{predClust}{A \code{cellpainting_clusters} object.}

\item{data.norm}{A \code{cellpainting_normalized} object.}

\item{method}{Statistical test. Currently only \code{"wilcox"} is supported.}

\item{p.adjust}{Multiple-testing correction method.}

\item{cutoff}{Adjusted p-value cutoff.}

\item{medianDiffCutoff}{Median-difference cutoff.}
}
\value{
A tibble of pairwise cluster feature statistics.
}
\description{
Select cluster-discriminating features among unknown clusters
}
